pyproject.toml
[project]
name = "blaue-tonne-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "camelot-py>=1.0.9",
    "fastapi[standard]>=0.117.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
]

[tool.ruff]
line-length = 120
